refactor(sdk): move generate_keypair to client-core
Extract ML-KEM keypair generation from rust-sdk into client-core
so all P0 operations for RFC-006 (UniFFI mobile targets) are
available in the platform-agnostic crate.
